Fix inherit base classes without virtual destructors
Several classes might be inherited in end-web-applications, but these
haven't virtual destructors, which might lead to memory leaks.

* Add virtual destruct the bellow classes:
  + Manager_base
  + Mailer
  + Email_base; Email
  + Poll
  + Socket; SocketGroup
  + Transceiver
  + Connection
  + Curler
  + FcgiStreambuf

Testing: added `virtual` to the `EmailSender::response()` method of
`email.cpp` example for tesing reasons; build the examples targets with
`-Wnon-virtual-dtor` flag, and there are has no warnings.

Please, note, the `-Wnon-virtual-dtor` flag is not added to the
CMakeLists to be consistent for end-applications, which already uses the
`fastcgipp`.

End-user-impact: None.

Signed-off-by: Igor Kononenko <i.kononenko@yadro.com>